rename research assistant to experimenter
