+++ metagpt/ext/sela/README.md
@@ -213,7 +213,7 @@ Run the following script to get the running results, a `log` folder and a `works
 The `log` folder will contain the experimental configuration and the generated scheme, and the `workspace` folder will save the final results generated by aide
 
 ```
-python experimenter/aide.py
+python runner/aide.py
 ```
 
 ### Autogluon
